requirements.txt
python-dotenv
lark
sympy
google-genai>=1.0.0
pydot
graphviz
fastapi
uvicorn
pydantic
python-multipart